worker: move migrations
